Relocate and simplify the Room actions
Those global actions were living spread between the no longer available
Sidebar and Button components.
