--- src/components/Button/Button.css
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) [2015-2019] SUSE Linux
- *
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE.txt file for details.
- */
-
-.Button {
-  background-color: #272727;
-  color: #fff;
-  display: inline-flex;
-  align-items: center;
-  width: 50px;
-  height: 50px;
-  border-bottom: 1px solid #575757;
-}
-
-.Button.red {
-  background-color: red;
-}
-
-.Button > a {
-  margin: 0;
-  height: 100%;
-  width: 100%;
-  line-height: 60px;
-  text-align: center;
-  font-size: 35px;
-  cursor: pointer;
-}
-
-.Button p {
-  margin: 0;
-}
