--- src/components/Button/index.js
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) [2015-2019] SUSE Linux
- *
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE.txt file for details.
- */
-
-export { default } from './Button';
