+++ src/components/Room/Actions/LogOut.test.js
@@ -1,14 +1,14 @@
 /**
- * Copyright (c) [2015-2019] SUSE Linux
+ * Copyright (c) [2020] SUSE Linux
  *
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE.txt file for details.
  */
 
 import React from 'react';
-import { renderWithRedux } from '../../setupTests';
-import Button from './Button';
+import LogOut from './LogOut';
+import { renderWithRedux } from '../../../setupTests';
 
 it('renders without crashing', () => {
-  renderWithRedux(<Button />);
+  renderWithRedux(<LogOut />);
 });
