+++ src/components/Room/Actions/ShareScreen.test.js
@@ -1,14 +1,14 @@
 /**
- * Copyright (c) [2015-2019] SUSE Linux
+ * Copyright (c) [2020] SUSE Linux
  *
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE.txt file for details.
  */
 
 import React from 'react';
-import { renderWithRedux } from '../../setupTests';
-import Sidebar from './Sidebar';
+import ShareScreen from './ShareScreen';
+import { renderWithRedux } from '../../../setupTests';
 
 it('renders without crashing', () => {
-  renderWithRedux(<Sidebar />);
+  renderWithRedux(<ShareScreen />);
 });
