+++ src/components/Room/Actions/index.js
@@ -1,8 +1,9 @@
 /**
- * Copyright (c) [2015-2019] SUSE Linux
+ * Copyright (c) [2020] SUSE Linux
  *
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE.txt file for details.
  */
 
-export { default } from './Sidebar';
+export { default as LogOut } from './LogOut';
+export { default as ShareScreen } from './ShareScreen';
