--- src/components/Sidebar/Sidebar.css
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) [2015-2019] SUSE Linux
- *
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE.txt file for details.
- */
-
-.Sidebar {
-  background-color: #272727;
-  box-sizing: border-box;
-  color: #fff;
-  grid-area: sidebar;
-}
